Deleted the tests associated with the stenn site.
