--- framework/website_stenn_page/articles_page.py
@@ -1,7 +1,0 @@
-from selenium.webdriver.common.by import By
-from base.base_page import BasePage
-
-
-class Articles(BasePage):
-    def get_heading_label(self):
-        return self.find_element((By.XPATH, '//h1[text()="Articles"]'))
