--- framework/website_stenn_page/factoring_works_page.py
@@ -1,8 +1,0 @@
-from selenium.webdriver.common.by import By
-from base.base_page import BasePage
-
-
-class FactoringWorks(BasePage):
-    def get_heading_label(self):
-        return self.find_element((By.XPATH, '//h1[contains(., "Everything You Need")]'))
-
