--- framework/website_stenn_page/main_page.py
@@ -1,4 +1,0 @@
-class MainPage:
-    def get_heading_label(self, driver):
-        # Ваш код для выполнения операций с driver
-        pass
