--- framework/website_stenn_page/questions_answers_page.py
@@ -1,8 +1,0 @@
-from selenium.webdriver.common.by import By
-from base.base_page import BasePage
-
-
-class QuestionsAnswers(BasePage):
-    def get_heading_label(self):
-        return self.find_element((By.XPATH, '//h2[text()="Questions & Answers"]'))
-
